#46 共通コンポーネントを１つにまとめた
