+++ src/components/templates/StoreManagementTemplate/index.tsx
@@ -1,27 +1,13 @@
 import { FC } from 'react';
-import { Box } from '@mui/system';
 
-import Header from '../../organisms/Header';
 import EmployeeListView from '../../organisms/EmployeeListView';
-import Sidebar from '../../organisms/Sidebar';
-
-const mainStyle = {
-  flexGrow: 1,
-  p: 3,
-  marginTop: '64px',
-};
+import StoreCommonComponent from '../../organisms/StoreCommonComponent';
 
 const Index: FC = () => (
   <>
-    <Box sx={{ display: 'flex' }}>
-      <header>
-        <Header />
-      </header>
-      <Sidebar />
-      <Box component="main" sx={mainStyle}>
-        <EmployeeListView />
-      </Box>
-    </Box>
+    <StoreCommonComponent>
+      <EmployeeListView />
+    </StoreCommonComponent>
   </>
 );
